requirements.txt
TgCrypto
ffmpeg-python
psutil
Pyrogram
pytgcalls==0.0.23
wheel
youtube_dl


